Move all_sites() from util to site
